healthcare_mcp/requirements.txt
mcp
psycopg2-binary
sentence-transformers
python-dotenv
numpy
